+++ kipper/core/src/compiler/ast/nodes/declarations/function-declaration/function-declaration.ts
@@ -10,8 +10,7 @@ import type { CompilableNodeParent } from "../../../compilable-ast-node";
 import type { CompoundStatement, Statement } from "../../statements";
 import type { IdentifierTypeSpecifierExpression } from "../../expressions";
 import type { RawType, ScopeFunctionDeclaration } from "../../../../semantics";
-import { BuiltInTypeFunc } from "../../../../semantics";
-import { FunctionScope } from "../../../../semantics";
+import { BuiltInTypeFunc, FunctionScope } from "../../../../semantics";
 import type { FunctionDeclarationContext } from "../../../../lexer-parser";
 import {
 	CompoundStatementContext,