+++ kipper/core/src/compiler/semantics/types/base/built-in-type.ts
@@ -1,7 +1,9 @@
 import { ProcessedType } from "./processed-type";
 import type { KipperBuiltInTypeLiteral } from "../../../const";
+import type { TypeError } from "../../../../errors";
 import { ArgumentAssignmentTypeError, AssignmentTypeError, PropertyAssignmentTypeError } from "../../../../errors";
 import { BuiltInTypes } from "../../symbol-table";
+import { UnionType } from "./union-type";
 
 /**
  * Represents a built-in type that is used in the type analysis phase.
@@ -35,19 +37,28 @@ export abstract class BuiltInType extends ProcessedType {
 	 * @since 0.11.0
 	 */
 	public assertAssignableTo(type: ProcessedType, propertyName?: string, argumentName?: string): void {
+		let e: TypeError | undefined = undefined;
 		if (this === type || type === BuiltInTypes.any) {
 			return;
+		} else if (type instanceof UnionType) {
+			if (type.unionTypes.some((unionType: ProcessedType) => this.isAssignableTo(unionType))) {
+				return;
+			} else {
+				e = new AssignmentTypeError(type.identifier, this.identifier);
+			}
 		} else if (
 			BuiltInType.interchangeableTypes.includes(this.identifier) &&
 			BuiltInType.interchangeableTypes.includes(type.identifier)
 		) {
 			return;
-		} else if (propertyName) {
-			throw new PropertyAssignmentTypeError(propertyName, type.identifier, this.identifier);
+		}
+
+		if (propertyName) {
+			throw new PropertyAssignmentTypeError(propertyName, type.identifier, this.identifier, e);
 		} else if (argumentName) {
-			throw new ArgumentAssignmentTypeError(argumentName, type.identifier, this.identifier);
+			throw new ArgumentAssignmentTypeError(argumentName, type.identifier, this.identifier, e);
 		} else {
-			throw new AssignmentTypeError(type.identifier, this.identifier);
+			throw new AssignmentTypeError(type.identifier, this.identifier, e);
 		}
 	}
 }