+++ kipper/core/src/compiler/semantics/types/built-in/index.ts
@@ -2,13 +2,14 @@
  * All built-in types in the Kipper language.
  * @since 0.12.0
  */
-export * from "./array";
+export * from "./any";
+export * from "./void";
+export * from "./undefined";
 export * from "./bool";
-export * from "./func";
 export * from "./null";
 export * from "./num";
 export * from "./str";
 export * from "./type";
-export * from "./undefined";
-export * from "./void";
+export * from "./func";
+export * from "./array";
 export * from "./obj";