+++ CHANGELOG.md
@@ -33,9 +33,11 @@ To use development versions of Kipper download the
 - Implemented the generic `Func<T..., R>` type and function type initializers.
   ([#584](https://github.com/Kipper-Lang/Kipper/issues/584))
 - Implemented internal generic spread argument `T...`, which allows multiple arguments to be passed to a single
-  parameter inside of a generic type specifier.
+  parameter inside a generic type specifier.
 - Implemented constant `NaN`, which represents the `NaN` value in JavaScript (Not a Number).
   ([#671](https://github.com/Kipper-Lang/Kipper/issues/671))
+- Support for internal type unions in built-in and internal functions.
+  ([#496](https://github.com/Kipper-Lang/Kipper/issues/496))
 - New module:
   - `semantics/runtime-built-ins`, which contains runtime built-in functions, variables and types.
   - `semantics/runtime-internals`, which contains the runtime internal functions.
@@ -91,7 +93,9 @@ To use development versions of Kipper download the
 
 ### Changed
 
-- Changed type from interface to class:
+- Argument type of built-in function `print` from `str` to `any`.
+  ([#667](https://github.com/Kipper-Lang/Kipper/issues/667))
+- Type from interface to class:
   - `InternalFunction`, which represents an internal function.
   - `BuiltInFunction`, which represents a built-in function.
   - `InternalFunctionArgument`, which represents an internal function argument.